Gen-AI-Projects/PDF-Summarization-Tool/requirements.txt
streamlit
pypdf
openai
sentence-transformers
faiss-cpu
langchain
langchain-huggingface
langchain-text-splitters
langchain-community
langchain-openai
python-dotenv